+++ muxers/mplex/tests/two_peers.rs
@@ -19,9 +19,9 @@
 // DEALINGS IN THE SOFTWARE.
 
 use futures::{channel::oneshot, prelude::*};
-use libp2p_core::muxing::StreamMuxerExt;
-use libp2p_core::{upgrade, Transport};
-use libp2p_tcp::TcpTransport;
+use libp2p::core::muxing::StreamMuxerExt;
+use libp2p::core::{upgrade, Transport};
+use libp2p::tcp::TcpTransport;
 
 #[test]
 fn client_to_server_outbound() {