+++ protocols/gossipsub/src/protocol.rs
@@ -572,13 +572,12 @@ mod tests {
     use crate::IdentTopic as Topic;
     use libp2p_core::identity::Keypair;
     use quickcheck::*;
-    use rand::Rng;
 
     #[derive(Clone, Debug)]
     struct Message(RawGossipsubMessage);
 
     impl Arbitrary for Message {
-        fn arbitrary<G: Gen>(g: &mut G) -> Self {
+        fn arbitrary(g: &mut Gen) -> Self {
             let keypair = TestKeypair::arbitrary(g);
 
             // generate an arbitrary GossipsubMessage using the behaviour signing functionality
@@ -588,8 +587,8 @@ mod tests {
                 config,
             )
             .unwrap();
-            let data = (0..g.gen_range(10, 10024))
-                .map(|_| g.gen())
+            let data = (0..g.gen_range(10..10024u32))
+                .map(|_| u8::arbitrary(g))
                 .collect::<Vec<_>>();
             let topic_id = TopicId::arbitrary(g).0;
             Message(gs.build_raw_message(topic_id, data).unwrap())
@@ -600,9 +599,9 @@ mod tests {
     struct TopicId(TopicHash);
 
     impl Arbitrary for TopicId {
-        fn arbitrary<G: Gen>(g: &mut G) -> Self {
-            let topic_string: String = (0..g.gen_range(20, 1024))
-                .map(|_| g.gen::<char>())
+        fn arbitrary(g: &mut Gen) -> Self {
+            let topic_string: String = (0..g.gen_range(20..1024u32))
+                .map(|_| char::arbitrary(g))
                 .collect::<String>()
                 .into();
             TopicId(Topic::new(topic_string).into())
@@ -613,8 +612,9 @@ mod tests {
     struct TestKeypair(Keypair);
 
     impl Arbitrary for TestKeypair {
-        fn arbitrary<G: Gen>(g: &mut G) -> Self {
-            let keypair = if g.gen() {
+        #[cfg(feature = "rsa")]
+        fn arbitrary(g: &mut Gen) -> Self {
+            let keypair = if bool::arbitrary(g) {
                 // Small enough to be inlined.
                 Keypair::generate_ed25519()
             } else {
@@ -623,6 +623,12 @@ mod tests {
                 Keypair::rsa_from_pkcs8(&mut rsa_key).unwrap()
             };
             TestKeypair(keypair)
+        }
+
+        #[cfg(not(feature = "rsa"))]
+        fn arbitrary(_g: &mut Gen) -> Self {
+            // Small enough to be inlined.
+            TestKeypair(Keypair::generate_ed25519())
         }
     }