+++ swarm-derive/CHANGELOG.md
@@ -1,6 +1,26 @@
-# 0.30.0 - [unreleased]
+# 0.30.1 [unreleased]
 
-- Remove support for removed `NetworkBehaviourEventProcess`.
+- Fix an issue where the derive would generate bad code if the type parameters between the behaviour and a custom
+  out event differed. See [PR 2907].
+- Fix an issue where the derive would generate incorrect code depending on available imports. See [PR 2921].
+
+[PR 2907]: https://github.com/libp2p/rust-libp2p/pull/2907
+[PR 2921]: https://github.com/libp2p/rust-libp2p/pull/2921
+
+# 0.30.0
+
+- Remove support for removed `NetworkBehaviourEventProcess`. See [PR 2840].
+
+- Remove support for custom `poll` method on `NetworkBehaviour` via `#[behaviour(poll_method =
+  "poll")]`. See [PR 2841].
+
+[PR 2840]: https://github.com/libp2p/rust-libp2p/pull/2840
+[PR 2841]: https://github.com/libp2p/rust-libp2p/pull/2841
+
+- Remove support for non-`NetworkBehaviour` fields on main `struct` via `#[behaviour(ignore)]`. See
+  [PR 2842].
+
+[PR 2842]: https://github.com/libp2p/rust-libp2p/pull/2842
 
 # 0.29.0